+++ publish/CHANGELOG.md
@@ -2,6 +2,8 @@
   - Adds spawn specific triggerRules to the field `spawn`.
   - Adds new fields `addItems` and `removeItems` to more easily add or remove items from containers.
   - Adds support for filtering with the data type items.
+  - Adds new parameters "par5", "par6", "par7", "par8" and "par9".
+  - Adds new parameters "item_*" to get the amount of a specific item in containers.
   - Fixes global key filters not being converted to lower case.
   - Fixes nested parameters not working.
   - Fixes zdo based parameters not working when spawning objects.